commit to address https://github.com/CBIIT/R-cometsAnalytics/issues/58
needed to fix duplicity in the columns and row vars specified if metabolites are both in exposure and outcome specification. thus we had to change specification of var position to actual variable names. also added check near zero variance for adjusted as well as exposure variables in addition to outcomes